core: various code cleanups to reduce the number of warnings and refactoring hints
- use collection expressions
- remove redundant type prefix (object.ReferenceEquals, ...)
- remove redundant generic types
- fix some types in comments, ...

+++ Doxense.Core/IO/Hashing/xxHash64.cs
@@ -396,7 +396,7 @@ namespace Doxense.IO.Hashing
 				h64 = state.Seed + PRIME64_5;
 			}
 
-			h64 += (ulong) totalLength;
+			h64 += totalLength;
 
 			while (p <= bEnd - 8)
 			{
